Ammended annotation in relation to fixed rt instruction
The reason rt was causing issues is because after a return, the cea would be at the address of the operand of the call instruction, not the address afterwards. This would try to run the operand of the call instruction as an instruction. This is in contrast to "trying to run the call again" as was thought to be the issue before. The code annotation has been updated to reflect this corrected reasoning.