Improve annotation quality; fix fvmr memory bug
Improve the annotation in the header files for the lexer, parser, and shared stages of fvma.
FVMR previously allocated twice as much memory initially as it was supposed to. This was due to dividing the number of bytes by 4 instead of 8 when calculating the size of the ROM. This has been updated to divide by eight, and to only add 1 to the size when necessary instead of always.